Refactor login form to distinct route. Add link to reset password (not yet implemented).
